Completed refactoring of ZoweJobNode.unit.test.ts
